backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-auth-oauthlib
